requirements.txt
torch
torchvision
tensorboardX
tensorboard
tensorflow
easydict
pandas
numpy
scipy
pillow
matplotlib
opencv-python